Update build and CI config
Fix #102
