--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["es2015", "stage-0"],
-  "plugins": ["transform-es2015-modules-umd"]
-}
