+++ babel.config.js
@@ -1,0 +1,17 @@
+const presets = [
+  [
+    '@babel/env',
+    {
+      targets: {
+        edge: '17',
+        firefox: '60',
+        chrome: '67',
+        safari: '11.1',
+      },
+      useBuiltIns: 'usage',
+      corejs: 3,
+    },
+  ],
+];
+
+module.exports = { presets };
