+++ package.json
@@ -5,7 +5,7 @@
     "Tom Alexander <me@tomalexander.co.nz>",
     "Misha Wolfson <misha@ginkgobioworks.com>"
   ],
-  "version": "2.0.1",
+  "version": "2.0.2",
   "license": "MIT",
   "main": "./dist/regression.js",
   "repository": "Tom-Alexander/regression-js",
@@ -18,25 +18,27 @@
   ],
   "scripts": {
     "clean": "rm -rf dist && mkdir dist",
-    "lint": "./node_modules/.bin/eslint src/** tests/**",
-    "test": "npm run lint && ./node_modules/.bin/nyc --reporter=lcov ./node_modules/.bin/mocha --compilers js:babel-core/register",
+    "lint": "eslint src/** test/**",
+    "test": "npm run lint && nyc --reporter=lcov mocha --require @babel/register",
     "build": "npm run clean && npm run build-regression-js && npm run build-regression-min-js",
-    "build-regression-js": "BABEL_ENV=production ./node_modules/.bin/babel src/regression.js --out-file dist/regression.js",
-    "build-regression-min-js": " ./node_modules/.bin/uglifyjs dist/regression.js --output dist/regression.min.js"
+    "build-regression-js": "BABEL_ENV=production babel  --presets=@babel/env src --out-dir dist",
+    "build-regression-min-js": "uglifyjs dist/regression.js --output dist/regression.min.js"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.25.0",
-    "babel-plugin-transform-es2015-modules-umd": "^6.24.1",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "chai": "^3.5.0",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb-base": "^11.2.0",
-    "eslint-plugin-import": "^2.7.0",
-    "mocha": "^3.2.0",
-    "nyc": "^11.0.3",
-    "uglify-js": "3"
+    "@babel/cli": "^7.7.7",
+    "@babel/core": "^7.7.7",
+    "@babel/preset-env": "^7.7.7",
+    "@babel/register": "^7.7.7",
+    "chai": "^4.2.0",
+    "eslint": "^6.8.0",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.19.1",
+    "mocha": "^6.2.2",
+    "nyc": "^15.0.0",
+    "uglify-js": "^3.7.3"
   },
-  "author": "Tom Alexander <me@tomalexander.co.nz>"
+  "author": "Tom Alexander <me@tomalexander.co.nz>",
+  "dependencies": {
+    "core-js": "3"
+  }
 }
