+++ src/regression.js
@@ -1,41 +1,6 @@
+const { round, deriveDataProperties } = require('./utils');
+
 const DEFAULT_OPTIONS = { order: 2, precision: 2, period: null };
-
-/**
-* Determine the coefficient of determination (r^2) of a fit from the observations
-* and predictions.
-*
-* @param {Array<Array<number>>} data - Pairs of observed x-y values
-* @param {Array<Array<number>>} results - Pairs of observed predicted x-y values
-*
-* @return {number} - The r^2 value, or NaN if one cannot be calculated.
-*/
-function determinationCoefficient(data, results) {
-  const predictions = [];
-  const observations = [];
-
-  data.forEach((d, i) => {
-    if (d[1] !== null) {
-      observations.push(d);
-      predictions.push(results[i]);
-    }
-  });
-
-  const sum = observations.reduce((a, observation) => a + observation[1], 0);
-  const mean = sum / observations.length;
-
-  const ssyy = observations.reduce((a, observation) => {
-    const difference = observation[1] - mean;
-    return a + (difference * difference);
-  }, 0);
-
-  const sse = observations.reduce((accum, observation, index) => {
-    const prediction = predictions[index];
-    const residual = observation[1] - prediction[1];
-    return accum + (residual * residual);
-  }, 0);
-
-  return 1 - (sse / ssyy);
-}
 
 /**
 * Determine the solution of a system of linear equations A * x = b using
@@ -85,21 +50,6 @@ function gaussianElimination(input, order) {
 }
 
 /**
-* Round a number to a precision, specificed in number of decimal places
-*
-* @param {number} number - The number to round
-* @param {number} precision - The number of decimal places to round to:
-*                             > 0 means decimals, < 0 means powers of 10
-*
-*
-* @return {numbr} - The number, rounded
-*/
-function round(number, precision) {
-  const factor = 10 ** precision;
-  return Math.round(number * factor) / factor;
-}
-
-/**
 * The set of all fitting methods
 *
 * @namespace
@@ -125,19 +75,16 @@ const methods = {
     const gradient = run === 0 ? 0 : round(rise / run, options.precision);
     const intercept = round((sum[1] / len) - ((gradient * sum[0]) / len), options.precision);
 
-    const predict = x => ([
+    const predict = (x) => ([
       round(x, options.precision),
       round((gradient * x) + intercept, options.precision)]
     );
 
-    const points = data.map(point => predict(point[0]));
-
     return {
-      points,
       predict,
       equation: [gradient, intercept],
-      r2: round(determinationCoefficient(data, points), options.precision),
       string: intercept === 0 ? `y = ${gradient}x` : `y = ${gradient}x + ${intercept}`,
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -160,19 +107,16 @@ const methods = {
     const b = ((sum[1] * sum[4]) - (sum[5] * sum[3])) / denominator;
     const coeffA = round(a, options.precision);
     const coeffB = round(b, options.precision);
-    const predict = x => ([
+    const predict = (x) => ([
       round(x, options.precision),
       round(coeffA * Math.exp(coeffB * x), options.precision),
     ]);
 
-    const points = data.map(point => predict(point[0]));
-
     return {
-      points,
       predict,
       equation: [coeffA, coeffB],
       string: `y = ${coeffA}e^(${coeffB}x)`,
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -193,19 +137,16 @@ const methods = {
     const coeffB = round(a, options.precision);
     const coeffA = round((sum[2] - (coeffB * sum[0])) / len, options.precision);
 
-    const predict = x => ([
+    const predict = (x) => ([
       round(x, options.precision),
       round(round(coeffA + (coeffB * Math.log(x)), options.precision), options.precision),
     ]);
 
-    const points = data.map(point => predict(point[0]));
-
     return {
-      points,
       predict,
       equation: [coeffA, coeffB],
       string: `y = ${coeffA} + ${coeffB} ln(x)`,
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -227,19 +168,16 @@ const methods = {
     const coeffA = round(Math.exp(a), options.precision);
     const coeffB = round(b, options.precision);
 
-    const predict = x => ([
+    const predict = (x) => ([
       round(x, options.precision),
       round(round(coeffA * (x ** coeffB), options.precision), options.precision),
     ]);
 
-    const points = data.map(point => predict(point[0]));
-
     return {
-      points,
       predict,
       equation: [coeffA, coeffB],
       string: `y = ${coeffA}x^${coeffB}`,
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 
@@ -275,17 +213,15 @@ const methods = {
     }
     rhs.push(lhs);
 
-    const coefficients = gaussianElimination(rhs, k).map(v => round(v, options.precision));
+    const coefficients = gaussianElimination(rhs, k).map((v) => round(v, options.precision));
 
-    const predict = x => ([
+    const predict = (x) => ([
       round(x, options.precision),
       round(
         coefficients.reduce((sum, coeff, power) => sum + (coeff * (x ** power)), 0),
         options.precision,
       ),
     ]);
-
-    const points = data.map(point => predict(point[0]));
 
     let string = 'y = ';
     for (let i = coefficients.length - 1; i >= 0; i--) {
@@ -300,10 +236,9 @@ const methods = {
 
     return {
       string,
-      points,
       predict,
       equation: [...coefficients].reverse(),
-      r2: round(determinationCoefficient(data, points), options.precision),
+      ...deriveDataProperties(data, predict, options),
     };
   },
 };
