+++ test/data.js
@@ -208,5 +208,4 @@ export const polynomial = {
     data: [[1, 6], [2, 11], [3, 18], [4, 27], [5, 38], [6, 51]],
     points: [[1, 6], [2, 11], [3, 18], [4, 27], [5, 38], [6, 51]],
   },
-
 };
